moving the RBFLayer, RBFNetwork to networks
